- Add ProjectTitleComponent.vue - Add ProjectStateComponent.vue - Fix is_group for projects - Add ProjectGroupComponent.vue - Fix link to project group - Add ProjectTeamComponent.vue - Add ProjectAttributesComponent.vue - Add RelevantDatesForShiftPlanningComponent.vue - Add ShiftContactPersonsComponent.vue - Add GeneralShiftInformationComponent.vue - Add ProjectBudgetDeadlineComponent.vue - Add ProjectDocumentsComponent.vue
