+++ lang/de.json
@@ -1669,5 +1669,6 @@
     "Shift start date": "Schichtbeginn",
     "Shift end date": "Schichtende",
     "The shift end time cannot be before the shift start time.": "Die Schichtendzeit kann nicht vor der Schichtstartzeit liegen.",
-    "Edit Key Visual": "Key Visual bearbeiten"
+    "Edit Key Visual": "Key Visual bearbeiten",
+    "No budget deadline specified": "Kein Budget-Stichtag angegeben"
 }
