+++ lang/en.json
@@ -1668,5 +1668,6 @@
     "Shift start date": "Shift start date",
     "Shift end date": "Shift end date",
     "The shift end time cannot be before the shift start time.": "The shift end time cannot be before the shift start time.",
-    "Edit Key Visual": "Edit Key Visual"
+    "Edit Key Visual": "Edit Key Visual",
+    "No budget deadline specified": "No budget deadline specified"
 }
