+++ resources/js/Pages/Projects/Components/ProjectBudgetDeadlineComponent.vue
@@ -1,0 +1,15 @@
+<template>
+    <span>
+        {{ $t('Budget deadline') }}: {{ project.budget_deadline ?? $t('No budget deadline specified') }}
+    </span>
+</template>
+
+<script>
+import {defineComponent} from "vue";
+
+export default defineComponent({
+    props: [
+        'project'
+    ]
+});
+</script>
